+++ example/lib/examples/mario_movement_example.dart
@@ -3,16 +3,9 @@ import 'package:flutter/services.dart';
 import 'package:tremble/tremble.dart';
 import 'package:tremble_example/tex_enum.dart';
 
-enum MarioState { idle, run, jump }
-
 class MarioMovementExample extends ScreenController {
   late final SpriteBatch batch;
-  late final Animation<Tex> mario;
-
-  final Vec2 acc = Vec2(0, 900);
-  final Vec2 vel = Vec2.zero();
-  final state = StateMachine<MarioState>(MarioState.jump);
-  final keys = <LogicalKeyboardKey>{};
+  late final Mario mario;
 
   @override
   Future<void> preload(UpdateCallback progress, VoidCallback done) async {
@@ -22,25 +15,67 @@ class MarioMovementExample extends ScreenController {
 
   @override
   void setup(BuildContext context, double width, double height) {
-    mario = Animation<Tex>(
-      animations: [
-        batch.getAnimation(Tex.marioBigJump, speed: 0),
-        batch.getAnimation(Tex.marioBigIdle, speed: 0),
-        batch.getAnimation(Tex.marioBigRun, speed: 8),
-        batch.getAnimation(Tex.marioBigSlide, speed: 0),
-      ],
-      position: Vec2(width * 0.5, height * 0.5),
-      scale: 1.5,
-    );
+    mario = Mario(batch: batch, x: width * 0.5, y: height * 0.5);
+  }
 
+  @override
+  void update(double deltaTime) {
+    mario.update(deltaTime);
+    mario.constraintBottom();
+  }
+
+  @override
+  void draw(Canvas canvas, Size size) {
+    batch.draw(canvas, [mario.anim]);
+    canvas.drawLine(const Offset(0, 385), Offset(size.width, 385), Paint()..color = Colors.white);
+  }
+
+  @override
+  void keyDown(LogicalKeyboardKey key) {
+    mario.keys.add(key);
+  }
+
+  @override
+  void keyUp(LogicalKeyboardKey key) {
+    mario.keys.remove(key);
+  }
+
+  @override
+  void dispose() {
+    batch.dispose();
+  }
+}
+
+//////////// MARIO CLASS
+enum MarioState { idle, run, jump }
+
+class Mario {
+  final Animation<Tex> anim;
+
+  final Vec2 acc = Vec2(0, 900);
+  final Vec2 vel = Vec2.zero();
+  final state = StateMachine<MarioState>(MarioState.jump);
+  final keys = <LogicalKeyboardKey>{};
+
+  Mario({required SpriteBatch batch, required double x, required double y})
+      : anim = Animation<Tex>(
+          animations: [
+            batch.getAnimation(Tex.marioBigJump, speed: 0),
+            batch.getAnimation(Tex.marioBigIdle, speed: 0),
+            batch.getAnimation(Tex.marioBigRun, speed: 8),
+            batch.getAnimation(Tex.marioBigSlide, speed: 0),
+          ],
+          position: Vec2(x, y),
+          scale: 1.5,
+        ) {
     state.register(
       MarioState.idle,
       onEnter: () {
-        mario.setAnimation(Tex.marioBigIdle);
+        anim.setAnimation(Tex.marioBigIdle);
         vel.x = 0;
       },
       onUpdate: (deltaTime) {
-        if (checkJump()) return;
+        if (_checkJump()) return;
 
         if (keys.contains(LogicalKeyboardKey.arrowRight) ||
             keys.contains(LogicalKeyboardKey.arrowLeft)) {
@@ -52,34 +87,35 @@ class MarioMovementExample extends ScreenController {
     state.register(
       MarioState.run,
       onEnter: () {
-        mario.setAnimation(Tex.marioBigRun, fromFrame: 0);
+        anim.setAnimation(Tex.marioBigRun, fromFrame: 0);
       },
       onUpdate: (deltaTime) {
-        if (checkJump()) return;
+        if (_checkJump()) return;
 
-        xMove();
+        _xMove();
 
         if (acc.x != 0) {
           if (vel.x != 0 && acc.x.sign != vel.x.sign) {
-            mario.setAnimation(Tex.marioBigSlide);
+            anim.setAnimation(Tex.marioBigSlide);
           } else {
-            mario.setAnimation(Tex.marioBigRun, fromFrame: 0);
+            anim.setAnimation(Tex.marioBigRun, fromFrame: 0);
           }
 
-          mario.flip = acc.x < 0;
+          anim.flip = acc.x < 0;
         } else if (vel.x.abs() < 20) {
           state.value = MarioState.idle;
         }
       },
     );
+
     state.register(
       MarioState.jump,
-      onEnter: () => mario.setAnimation(Tex.marioBigJump),
-      onUpdate: (deltaTime) => xMove(),
+      onEnter: () => anim.setAnimation(Tex.marioBigJump),
+      onUpdate: (deltaTime) => _xMove(),
     );
   }
 
-  void xMove() {
+  void _xMove() {
     acc.x = 0;
     if (keys.contains(LogicalKeyboardKey.arrowRight)) {
       acc.x += 400;
@@ -89,7 +125,7 @@ class MarioMovementExample extends ScreenController {
     }
   }
 
-  bool checkJump() {
+  bool _checkJump() {
     if (keys.contains(LogicalKeyboardKey.arrowUp)) {
       vel.y = -400;
       state.value = MarioState.jump;
@@ -98,21 +134,21 @@ class MarioMovementExample extends ScreenController {
     return false;
   }
 
-  @override
   void update(double deltaTime) {
     state.update(deltaTime);
 
     vel.add(acc * deltaTime);
-    mario.position.add(vel * deltaTime);
+    anim.position.add(vel * deltaTime);
 
-    mario.update(deltaTime);
+    anim.update(deltaTime);
 
     acc.x = 0;
     vel.x = MathUtils.damp(vel.x, 0, 2.5, deltaTime);
+  }
 
-    // Collide bottom
-    if (mario.position.y > 360 && vel.y >= 0) {
-      mario.position.y = 360;
+  void constraintBottom() {
+    if (anim.position.y > 360 && vel.y >= 0) {
+      anim.position.y = 360;
       vel.y = 0;
 
       if (state.value == MarioState.jump) {
@@ -123,26 +159,5 @@ class MarioMovementExample extends ScreenController {
         }
       }
     }
-  }
-
-  @override
-  void draw(Canvas canvas, Size size) {
-    batch.draw(canvas, [mario]);
-    canvas.drawLine(const Offset(0, 385), Offset(size.width, 385), Paint()..color = Colors.white);
-  }
-
-  @override
-  void keyDown(LogicalKeyboardKey key) {
-    keys.add(key);
-  }
-
-  @override
-  void keyUp(LogicalKeyboardKey key) {
-    keys.remove(key);
-  }
-
-  @override
-  void dispose() {
-    batch.dispose();
   }
 }